Keep remoteIP and remotePort per session
the remoteIP and remotePort of the UDP socket are nulled when no bytes
are available on ESP8266 (Arduino Ethernet 'remembers' last known IP and
port). In order to be save on multiple hardware, keep remoteIP and
remotePort per session.

corrected ESP6288 filename

sample for Arduino and ESP8266 only send notes when connected